agent_call_rework
