+++ evals/common/scoring.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 import json
 from collections import Counter
 from dataclasses import dataclass, field
+from itertools import pairwise
 from pathlib import Path
 
 # Row comparison lives in the trilogy package so the eval harness and the
@@ -63,6 +64,14 @@ class AgentMetrics:
     # Per-tool count of calls whose arguments matched a prior call in the same
     # query (signal: the agent is looping on the same exploration step).
     repeated_calls_by_name: dict[str, int] = field(default_factory=dict)
+    # Maximal runs of consecutive `trilogy agent-info ...` calls. Each item is
+    # the routed topic (`index`, `query`, `authoring`, `syntax example ...`).
+    # This exposes whether an agent followed the root directory into a useful
+    # drilldown before doing unrelated work.
+    agent_info_sequences: list[list[str]] = field(default_factory=list)
+    agent_info_transitions: dict[str, int] = field(default_factory=dict)
+    agent_info_directory_calls: int = 0
+    agent_info_directory_followups: int = 0
 
     @property
     def tool_success_rate(self) -> float:
@@ -121,6 +130,16 @@ def _call_signature(name: str, args: dict) -> str:
     return json.dumps({"name": name, "args": args}, sort_keys=True, default=str)
 
 
+def _agent_info_topic(name: str, args: dict) -> str | None:
+    if name != "trilogy":
+        return None
+    cli_args = args.get("args") or []
+    if not isinstance(cli_args, list) or not cli_args or cli_args[0] != "agent-info":
+        return None
+    parts = [str(part).strip() for part in cli_args[1:] if str(part).strip()]
+    return " ".join(parts) if parts else "index"
+
+
 def parse_agent_log(log_path: Path) -> AgentMetrics:
     """Aggregate the agent's ``--log-file`` JSONL trace into harness metrics."""
     m = AgentMetrics()
@@ -130,6 +149,14 @@ def parse_agent_log(log_path: Path) -> AgentMetrics:
     repeated: Counter[str] = Counter()
     seen_signatures: dict[str, str] = {}
     pending_call_name: list[tuple[str, str]] = []  # (name, signature) queue
+    agent_info_chain: list[str] = []
+
+    def flush_agent_info_chain() -> None:
+        if not agent_info_chain:
+            return
+        m.agent_info_sequences.append(list(agent_info_chain))
+        agent_info_chain.clear()
+
     if not log_path.exists():
         return m
 
@@ -154,17 +181,23 @@ def parse_agent_log(log_path: Path) -> AgentMetrics:
             by_name[name] += 1
             m.tool_calls_total += 1
             args = event.get("arguments") or {}
-            sig = _call_signature(name, args if isinstance(args, dict) else {})
+            safe_args = args if isinstance(args, dict) else {}
+            topic = _agent_info_topic(name, safe_args)
+            if topic is None:
+                flush_agent_info_chain()
+            else:
+                agent_info_chain.append(topic)
+            sig = _call_signature(name, safe_args)
             if sig in seen_signatures.values():
                 repeated[name] += 1
             seen_signatures[f"{name}#{m.tool_calls_total}"] = sig
             pending_call_name.append((name, sig))
             if name == "trilogy":
-                cli_args = args.get("args") or []
+                cli_args = safe_args.get("args") or []
                 if isinstance(cli_args, list) and cli_args:
                     subcommands[str(cli_args[0])] += 1
             elif name == "return_control_to_user":
-                msg = args.get("message")
+                msg = safe_args.get("message")
                 if isinstance(msg, str):
                     m.farewell = msg
         elif etype == "reviewer_verdict":
@@ -188,6 +221,17 @@ def parse_agent_log(log_path: Path) -> AgentMetrics:
             if pending_call_name:
                 pending_call_name.pop(0)
 
+    flush_agent_info_chain()
+    transitions: Counter[str] = Counter()
+    for sequence in m.agent_info_sequences:
+        for source, destination in pairwise(sequence):
+            transitions[f"{source} -> {destination}"] += 1
+        m.agent_info_directory_calls += sequence.count("index")
+        m.agent_info_directory_followups += sum(
+            source == "index" and destination != "index"
+            for source, destination in pairwise(sequence)
+        )
+    m.agent_info_transitions = dict(transitions)
     m.tool_calls_by_name = dict(by_name)
     m.trilogy_subcommands = dict(subcommands)
     m.tool_output_stats = stats
@@ -217,6 +261,9 @@ def aggregate_metrics(metrics_list: list[AgentMetrics]) -> AgentMetrics:
         by_name.update(m.tool_calls_by_name)
         subcmds.update(m.trilogy_subcommands)
         repeated.update(m.repeated_calls_by_name)
+        agg.agent_info_sequences.extend(m.agent_info_sequences)
+        agg.agent_info_directory_calls += m.agent_info_directory_calls
+        agg.agent_info_directory_followups += m.agent_info_directory_followups
         for tool, s in m.tool_output_stats.items():
             bucket = output_stats.setdefault(tool, ToolOutputStats())
             bucket.count += s.count
@@ -228,6 +275,10 @@ def aggregate_metrics(metrics_list: list[AgentMetrics]) -> AgentMetrics:
     agg.tool_calls_by_name = dict(by_name)
     agg.trilogy_subcommands = dict(subcmds)
     agg.repeated_calls_by_name = dict(repeated)
+    info_transitions: Counter[str] = Counter()
+    for m in metrics_list:
+        info_transitions.update(m.agent_info_transitions)
+    agg.agent_info_transitions = dict(info_transitions)
     agg.tool_output_stats = output_stats
     return agg
 
@@ -249,6 +300,10 @@ def metrics_to_dict(m: AgentMetrics) -> dict:
         "reviewer_kickbacks": m.reviewer_kickbacks,
         "farewell": m.farewell,
         "repeated_calls_by_name": dict(m.repeated_calls_by_name),
+        "agent_info_sequences": [list(sequence) for sequence in m.agent_info_sequences],
+        "agent_info_transitions": dict(m.agent_info_transitions),
+        "agent_info_directory_calls": m.agent_info_directory_calls,
+        "agent_info_directory_followups": m.agent_info_directory_followups,
         "tool_output_stats": {
             tool: {
                 "count": s.count,
@@ -277,6 +332,12 @@ def metrics_from_dict(d: dict) -> AgentMetrics:
         reviewer_kickbacks=d.get("reviewer_kickbacks", 0),
         farewell=d.get("farewell", ""),
         repeated_calls_by_name=dict(d.get("repeated_calls_by_name", {})),
+        agent_info_sequences=[
+            list(sequence) for sequence in d.get("agent_info_sequences", [])
+        ],
+        agent_info_transitions=dict(d.get("agent_info_transitions", {})),
+        agent_info_directory_calls=d.get("agent_info_directory_calls", 0),
+        agent_info_directory_followups=d.get("agent_info_directory_followups", 0),
         tool_output_stats={
             tool: ToolOutputStats(
                 count=s.get("count", 0),
