+++ trilogy/scripts/agent_info_docs/__init__.py
@@ -1,0 +1,1 @@
+"""Long-form documentation used by ``trilogy agent-info`` drilldowns."""
